refactor(#604): extract NavigationCoordinator from AppDelegate
Moves all view-factory methods (mainView, stepLogFromMain, settingsView,
runnerDetailView, scopeDetailView, validatedView) into a
NavigationCoordinator.swift extension on AppDelegate.
AppDelegate.swift retains only panel lifecycle, status icon,
and wiring.

Closes #604
